fixing warnings
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@26909 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ uvm-lib/api/com/untangle/uvm/node/PipelineEndpoints.java
@@ -96,7 +96,7 @@ public class PipelineEndpoints extends LogEvent
         // Begin & end and all in between have the same ID
         sessionId = begin.id();
 
-        protocol = (short)begin.protocol();
+        protocol = begin.protocol();
 
         cClientAddr = begin.clientAddr();
         cClientPort = begin.clientPort();
@@ -119,7 +119,7 @@ public class PipelineEndpoints extends LogEvent
     public PipelineEndpoints(IPSessionDesc begin)
     {
         sessionId = begin.id();
-        protocol = (short)begin.protocol();
+        protocol = begin.protocol();
         clientIntf = begin.clientIntf();
         // Don't fill in anything that can change later.
     }
